@@ -1,49 +1,88 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+import io.gitlab.arturbosch.detekt.Detekt
 
 plugins {
-	id("org.springframework.boot") version "2.2.1.BUILD-SNAPSHOT"
-	id("io.spring.dependency-management") version "1.0.8.RELEASE"
-	kotlin("jvm") version "1.3.50"
-	kotlin("plugin.spring") version "1.3.50"
-	id("org.jetbrains.kotlin.plugin.jpa") version "1.3.50" apply false
+	`build-dashboard`
+	id("com.dorongold.task-tree") version "3.0.0"
+	id("org.owasp.dependencycheck") version "9.2.0"
+	id("com.github.ben-manes.versions") version "0.51.0"
+	id("io.gitlab.arturbosch.detekt") version("1.23.6")
+	jacoco
+
+	id("org.springframework.boot") version "3.2.5"
+	id("io.spring.dependency-management") version "1.1.4"
+	kotlin("jvm") version "1.9.10"
+	kotlin("plugin.spring") version "1.9.10"
+	id("org.jetbrains.kotlin.plugin.jpa") version "1.9.10" apply false
 }
+
+//apply(plugin = "io.spring.dependency-management")
 
 allprojects {
 	group = "com.stringconcat"
-	version = "0.0.1-SNAPSHOT"
+	version = "0.0.1"
 
 	repositories {
 		mavenCentral()
-		jcenter()
 		maven { url = uri("https://repo.spring.io/milestone") }
 		maven { url = uri("https://repo.spring.io/snapshot") }
 	}
 
-	tasks.withType<KotlinCompile> {
-		kotlinOptions {
-			freeCompilerArgs = listOf("-Xjsr305=strict")
-			jvmTarget = "1.8"
+//	tasks.named("compileKotlin", org.jetbrains.kotlin.gradle.tasks.KotlinCompilationTask::class.java) {
+//		compilerOptions {
+//			freeCompilerArgs.add("-Xjsr305=strict")
+//		}
+//	}
+
+	apply(plugin = "kotlin")
+	apply(plugin = "io.gitlab.arturbosch.detekt")
+
+
+//	apply {
+//		plugin("org.springframework.boot")
+//		plugin("io.spring.dependency-management")
+//	}
+
+//	apply(plugin = "java")
+//	apply(plugin = "kotlin")
+//	apply(plugin = "io.spring.dependency-management")
+
+//	apply {
+//		plugin("io.gitlab.arturbosch.detekt")
+//	}
+
+	detekt {
+		repositories {
+			mavenCentral()
+		}
+
+		toolVersion = "1.23.3"
+		config.setFrom(file("${rootProject.projectDir}/tools/detekt/detekt-config.yml"))
+		buildUponDefaultConfig = true
+
+		source.from("src/main/kotlin", "src/test/kotlin", "src/test/gatling")
+
+		dependencies {
+			detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting:1.23.6")
+		}
+	}
+
+	tasks.withType<Detekt>().configureEach {
+		reports {
+			html.required.set(true)
 		}
 	}
 
 }
 
-java.sourceCompatibility = JavaVersion.VERSION_1_8
-
-val developmentOnly by configurations.creating
-configurations {
-	runtimeClasspath {
-		extendsFrom(developmentOnly)
-	}
-}
-
+java.sourceCompatibility = JavaVersion.VERSION_17
+java.targetCompatibility = JavaVersion.VERSION_17
 
 
 dependencies {
 	// spring modules
-	implementation("org.springframework.boot:spring-boot-starter-webflux")
-	implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-	implementation("org.springframework.boot:spring-boot-starter-data-rest")
+	implementation("org.springframework.boot:spring-boot-starter-webflux:2.7.2")
+	implementation("org.springframework.boot:spring-boot-starter-data-jpa:2.7.2")
+	implementation("org.springframework.boot:spring-boot-starter-data-rest:2.7.2")
 
 	// kotlin
 	implementation("org.jetbrains.kotlin:kotlin-reflect")
