@@ -2,7 +2,8 @@ plugins {
     kotlin("jvm")
 }
 
-java.sourceCompatibility = JavaVersion.VERSION_1_8
+java.sourceCompatibility = JavaVersion.VERSION_17
+java.targetCompatibility = JavaVersion.VERSION_17
 
 dependencies {
     implementation("javax.inject:javax.inject:1")
@@ -11,6 +12,6 @@ dependencies {
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
 
     // tests
-    testCompile("org.junit.jupiter:junit-jupiter-api:5.5.2")
+//    testCompile("org.junit.jupiter:junit-jupiter-api:5.5.2")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.5.2")
 }
