@@ -4,7 +4,8 @@ plugins {
     id("org.jetbrains.kotlin.plugin.jpa")
 }
 
-java.sourceCompatibility = JavaVersion.VERSION_1_8
+java.sourceCompatibility = JavaVersion.VERSION_17
+java.targetCompatibility = JavaVersion.VERSION_17
 
 dependencies {
     implementation(project(":businessPeople"))
@@ -12,13 +13,13 @@ dependencies {
 
     implementation("javax.persistence:javax.persistence-api:2.2")
     implementation("org.apache.logging.log4j:log4j-core:2.17.2")
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa:2.2.1.BUILD-SNAPSHOT")
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa:2.7.2")
 
     // kotlin
     implementation("org.jetbrains.kotlin:kotlin-reflect")
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
 
     // tests
-    testCompile("org.junit.jupiter:junit-jupiter-api:5.5.2")
+//    testCompile("org.junit.jupiter:junit-jupiter-api:5.5.2")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.5.2")
 }
