@@ -3,6 +3,7 @@ package com.stringconcat.people.useCasePeople
 import com.stringconcat.people.businessPeople.Person
 import com.stringconcat.people.businessPeople.PersonGenerator
 import java.time.LocalDate
+import java.util.*
 import javax.inject.Named
 
 @Named
@@ -13,7 +14,7 @@ class CreateNewPersonUseCase(
     operator fun invoke(
             personInput: PersonCreationSummary
     ): Person {
-        val inputSex = when(personInput.gender.toLowerCase()) {
+        val inputSex = when (personInput.gender.lowercase(Locale.getDefault())) {
             "male" -> Person.Sex.MAN
             "female" -> Person.Sex.WOMAN
             else -> Person.Sex.MAN
