@@ -1,12 +1,12 @@
 package com.stringconcat.people.presentation.controller
 
+import com.stringconcat.people.presentation.model.PersonRespectfullViewModel
+import com.stringconcat.people.presentation.view.personDetailsForm
+import com.stringconcat.people.presentation.view.renderDetailedView
 import com.stringconcat.people.useCasePeople.CreateNewPersonUseCase
 import com.stringconcat.people.useCasePeople.GetPersonUseCase
 import com.stringconcat.people.useCasePeople.MeUseCase
 import com.stringconcat.people.useCasePeople.PersonCreationSummary
-import com.stringconcat.people.presentation.model.PersonRespectfullViewModel
-import com.stringconcat.people.presentation.view.personDetailsForm
-import com.stringconcat.people.presentation.view.renderDetailedView
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
@@ -15,15 +15,14 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestMethod
 import org.springframework.web.bind.annotation.ResponseBody
-import java.io.*
 import java.net.URI
 import java.util.*
 
 @Controller
 class PeopleController(
-        val getPerson: GetPersonUseCase,
-        val createNew: CreateNewPersonUseCase,
-        val getMe: MeUseCase
+    val getPerson: GetPersonUseCase,
+    val createNew: CreateNewPersonUseCase,
+    val getMe: MeUseCase
 ) {
 
     @RequestMapping(value = ["/me"], method = [RequestMethod.GET])
@@ -37,15 +36,16 @@ class PeopleController(
         val idUUD = try {
             UUID.fromString(id)
         } catch (e: IllegalArgumentException) {
+            println(e.message)
             return ResponseEntity.badRequest().build()
         }
 
-        val person = getPerson(idUUD)
-                ?: return ResponseEntity.badRequest().build()
-
-        return ResponseEntity.ok(
+        return when (val person = getPerson(idUUD)) {
+            null -> ResponseEntity.badRequest().build()
+            else -> ResponseEntity.ok(
                 renderDetailedView(PersonRespectfullViewModel(person))
-        )
+            )
+        }
     }
 
     @RequestMapping(value = ["/generate"], method = [RequestMethod.GET])
@@ -55,16 +55,17 @@ class PeopleController(
     }
 
     @RequestMapping(
-            value = ["/generate"],
-            method = [RequestMethod.POST],
-            consumes = [MediaType.APPLICATION_FORM_URLENCODED_VALUE])
+        value = ["/generate"],
+        method = [RequestMethod.POST],
+        consumes = [MediaType.APPLICATION_FORM_URLENCODED_VALUE]
+    )
     @ResponseBody
-    fun create(personInput: PersonCreationSummary): ResponseEntity<String>{
+    fun create(personInput: PersonCreationSummary): ResponseEntity<String> {
         val generatedPerson = createNew(personInput)
 
         return ResponseEntity
-                .status(HttpStatus.FOUND)
-                .location(URI.create("/id/${generatedPerson.id}"))
-                .build()
+            .status(HttpStatus.FOUND)
+            .location(URI.create("/id/${generatedPerson.id}"))
+            .build()
     }
 }
