@@ -1,8 +1,17 @@
 package com.stringconcat.people.presentation.view
 
-import kotlinx.html.*
+import kotlinx.html.html
 import kotlinx.html.stream.appendHTML
 import java.lang.StringBuilder
+import kotlinx.html.FormMethod
+import kotlinx.html.body
+import kotlinx.html.dateInput
+import kotlinx.html.div
+import kotlinx.html.form
+import kotlinx.html.h1
+import kotlinx.html.p
+import kotlinx.html.submitInput
+import kotlinx.html.textInput
 
 fun personDetailsForm(): String =
     StringBuilder()
@@ -11,15 +20,32 @@ fun personDetailsForm(): String =
                 bootstrapHeader()
                 body {
                     div(classes = "header") {
-                        h1 {+ "Create a new person"}
+                        h1 {
+                            +"Create a new person"
+                        }
                     }
                     div("body") {
                         form(action = "/generate", method = FormMethod.post) {
-                            p { +"First name " }; textInput(name = "firstName")
-                            p { +"Last name " }; textInput(name = "secondName")
-                            p { +"Birthdate " }; dateInput(name = "birthDate")
-                            p { +"Gender " }; textInput(name = "gender")
-                            p { + "" }; submitInput()
+                            p {
+                                +"First name "
+                            }
+                            textInput(name = "firstName")
+                            p {
+                                +"Last name "
+                            }
+                            textInput(name = "secondName")
+                            p {
+                                +"Birthdate "
+                            }
+                            dateInput(name = "birthDate")
+                            p {
+                                +"Gender "
+                            }
+                            textInput(name = "gender")
+                            p {
+                                +""
+                            }
+                            submitInput()
                         }
                     }
                 }
