@@ -8,7 +8,7 @@ import org.springframework.web.client.ResponseErrorHandler
 import org.springframework.web.client.RestTemplate
 import javax.inject.Named
 
-const val defaultQuote = "whoops, something went wrong"
+const val DEFAULT_QUOTE = "whoops, something went wrong"
 
 @Named
 class QuoteGardenProvider : QuotesProvider {
@@ -16,17 +16,16 @@ class QuoteGardenProvider : QuotesProvider {
     private val getRandomUrl = "https://api.forismatic.com/api/1.0/?method=getQuote&format=json"
 
     override fun randomQuote(): Quote {
-
         val restTemplate = RestTemplate()
         restTemplate.errorHandler = EmptyErrorHandler()
 
         val response = restTemplate.getForEntity(getRandomUrl, QuoteResponse::class.java)
 
         check(response.statusCode == HttpStatus.OK) {
-            return defaultQuote
+            return DEFAULT_QUOTE
         }
 
-        return response.body?.quoteText ?: defaultQuote
+        return response.body?.quoteText ?: DEFAULT_QUOTE
     }
 
     internal data class QuoteResponse(
