@@ -9,13 +9,18 @@ import javax.inject.Named
 class MeUseCase(
         private val persistPerson: PersistPerson
 ) {
+    companion object {
+        private const val MOCK_YEAR_OF_BIRTH = 1987
+        private const val MOCK_MONTH_OF_BIRTH = 12
+        private const val MOCK_DAY_OF_BIRTH_MONTH = 1
+    }
+
     operator fun invoke(): Person {
-        val id = UUID.fromString("29f4d7e3-fd7c-4664-ad07-763326215ec4")
         val me = Person(
                 id = UUID.fromString("29f4d7e3-fd7c-4664-ad07-763326215ec4"),
                 firstName = "Sergey",
                 secondName = "Bukharov",
-                birthDate = LocalDate.of(1987,12,1),
+                birthDate = LocalDate.of(MOCK_YEAR_OF_BIRTH, MOCK_MONTH_OF_BIRTH, MOCK_DAY_OF_BIRTH_MONTH),
                 sex = Person.Sex.MAN,
                 avatartUrl = "https://avatars.dicebear.com/v2/male/my-somffething.svg",
                 favoriteQuote = "make the easy things easy, and the hard things possible"
